+++ zerofps/src/zerofps/render/mad.cpp
@@ -77,7 +77,7 @@ void MadInstans::Draw(void)
 
 	int iNumOfFrame = pkCore->GetAnimationTimeInFrames(iActiveAnimation);
 
-	int iFrame = fCurrentTime / 0.1;
+	int iFrame = int(fCurrentTime / 0.1);
